doc(README): Adds instructions for load_all()
And the instructions for Windows users if they get the dreaded file
too big error from the C++ compiling of TMB. For more information see
the following url
https://github.com/NOAA-FIMS/FIMS?tab=readme-ov-file#windows-users
where I learned the trick. Also, this has to be ran every time you open
up a new R session and expect to compile.